remove external custom setup
